New: checks if security group already exists

+++ scripts/module_02/helpers.js
@@ -1,4 +1,5 @@
-const fs = require('fs')
+const fs = require('fs');
+const fsPromises = fs.promises;
 const path = require('path')
 const os = require('os')
 
@@ -10,16 +11,29 @@ function persistKeyPair (keyData) {
       mode: 0o600
     }
 
+    if (fs.existsSync(keyPath)) {
+      resolve(keyData.KeyName);
+      return;
+    }
+
     fs.writeFile(keyPath, keyData.KeyMaterial, options, (err) => {
       if (err) reject(err)
       else {
         console.log('Key written to', keyPath)
-        resolve(keyData.KeyName)
+        resolve(keyData.KeyName);
+        return;
       }
     })
   })
 }
 
+function readKeyPair (keyName) {
+  const keyPath = path.join(os.homedir(), '.ssh', keyName);
+  const options = 'utf8';
+  return fsPromises.readFile(keyPath, options);
+}
+
 module.exports = {
-  persistKeyPair
+  persistKeyPair,
+  readKeyPair,
 }
